% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function to specific columns of a data frame}
\usage{
df_apply(.data, .fun, .filter, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{.fun}{A function that takes in a vector and returns a vector.}

\item{.filter}{A function that takes in a vector and returns a logical vector.}

\item{...}{Additional arguments to pass to .fun and .filter.}
}
\value{
A data frame that is a modified version of .data
}
\description{
Apply a function to specific columns of a data frame
}
\examples{
df <- data.frame(
  A = c(1, 2, 3),
  B = c(4, 5, 6),
  C = c("foo", "bar", "baz"))
my_fun <- function(x) x * 2
my_filter <- function(x) is.numeric(x)
df_apply(df, my_fun, my_filter)
}
