% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_sbuck.R
\name{diff_sbuck}
\alias{diff_sbuck}
\title{Calculation of difference of coffeehouses between two countries}
\usage{
diff_sbuck(code1, code2)
}
\arguments{
\item{code1}{Character. A two-letter combination code of first country}

\item{code2}{Character. A two-letter combination code of second country}
}
\value{
Numeric. A difference of total number Starbucks coffeehouses between two countries.
\itemize{
\item \code{difference} A difference value.
}
}
\description{
This function calculates the difference of total number of Starbucks coffeehouses between two countries,
by entering two-letter combination country codes of the countries.
}
\examples{
# Define Countries ISO country code:
# Define Australia's ISO country code:
# filter(ISOcode, Country == "Australia")[1,1]

# Calculate the difference between Australia and China.
# code1 = "AU"
# code2 = "CN"
diff_sbuck(code1 = "AU", code2 = "CN")


}
