% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dunkin_donuts}
\alias{dunkin_donuts}
\title{Dunkin' Donuts coffeehouses data}
\format{
A data frame with 4898 rows and 8 variables:
\describe{
\item{biz_name}{Store name of coffeehouse}
\item{e_address}{Street address of coffeehouse}
\item{e_city}{Name of city where coffeehouse located}
\item{e_state}{Name of state/province where coffeehouse located}
\item{e_postal}{Postcode}
\item{e_country}{Country where coffeehouse located}
\item{loc_LONG_poly}{Longitude}
\item{loc_LAT_poly}{Latitude}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2018/2018-05-07/week6_coffee_chains.xlsx}
}
\usage{
dunkin_donuts
}
\description{
A dataset containing Dunkin' Donuts coffeehouses' business information and
geographical location information, global wide.
}
\keyword{datasets}
