% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explosbuck-package.R
\docType{package}
\name{explosbuck-package}
\alias{explosbuck}
\alias{explosbuck-package}
\title{explosbuck: Starbucks Coffeehouses Business and Geographical Data}
\description{
An algorithm to explore Starbucks coffeehouses data .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2022/rpkg-xxuu0086}
  \item \url{https://etc5523-2022.github.io/rpkg-xxuu0086/}
  \item Report bugs at \url{https://github.com/etc5523-2022/rpkg-xxuu0086/issues}
}

}
\author{
\strong{Maintainer}: Xianghe Xu \email{xxuu0086@student.monash.edu}

}
\keyword{internal}
