% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISOcode}
\alias{ISOcode}
\title{ISO Country code}
\format{
A data frame with 65 rows and 2 variables:
\describe{
\item{Country_code}{A two letter combination code representing a country or an area}
\item{Country}{Name of a country}
}
}
\usage{
ISOcode
}
\description{
A dataset containing names of countries and ISO country codes.
}
\keyword{datasets}
