% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sbuck.R
\name{plot_sbuck}
\alias{plot_sbuck}
\title{Plot of the total Starbucks coffeehouses (Shiny-parameterized function)}
\usage{
plot_sbuck(code)
}
\arguments{
\item{code}{Character. Two-letter combination country codes}
}
\value{
\itemize{
\item \code{A plot}
}
}
\description{
This function plots the total number of Starbucks coffeehouses between countries,
by entering a two-letter combination country code of that country/countries.
}
\examples{

# Define Australia's ISO country code:
# filter(ISOcode, Country == "Australia")[1,1]

# Plot the countries for a comparison
# Australia code = "AU"
# United State code = "US"
# Canada code = "CA"
# China code = "CN"
plot_sbuck(c("AU", "US", "CA", "CN"))


}
