% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launching Shiny app}
\usage{
run_app()
}
\value{
Shiny app.
}
\description{
This function will launch the shiny app from the package.
Topic of the shiny app is about global coffee chains.
}
