% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starbucks}
\alias{starbucks}
\title{Starbucks coffeehouses data}
\format{
A data frame with 25600 rows and 13 variables:
\describe{
\item{Brand}{Coffeehouse brand}
\item{Store Number}{Store number of coffeehouse}
\item{Store Name}{Store name of coffeehouse}
\item{Ownership Type}{Company owned or licensed}
\item{Street Address}{Street address of coffeehouse}
\item{City}{Name of city where coffeehouse is located}
\item{State/Province}{Name of state/province where coffeehouse is located}
\item{Country}{Country where coffeehouse located}
\item{Postcode}{Postcode}
\item{Phone Number}{Phone Number}
\item{Timezone}{Timezone}
\item{Longitude}{Longitude of coffeehouse location}
\item{Latitude}{Latitude of coffeehouse location}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2018/2018-05-07/week6_coffee_chains.xlsx}
}
\usage{
starbucks
}
\description{
A dataset containing Starbucks coffeehouses' business information and
geographical location information, global wide.
}
\keyword{datasets}
