% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tim_hortons}
\alias{tim_hortons}
\title{Tim Hortons coffeehouses data}
\format{
A data frame with 4955 rows and 6 variables:
\describe{
\item{country}{Country where coffeehouse is located}
\item{address}{Street address of coffeehouse}
\item{city}{Name of city where coffeehouse is located}
\item{postal_code}{Postcode}
\item{state}{State where coffeehouse is located}
\item{store_name}{Name of coffeehouses}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2018/2018-05-07/week6_coffee_chains.xlsx}
}
\usage{
tim_hortons
}
\description{
A dataset containing Tim Hortons coffeehouses' business information and
geographical location information, global wide.
}
\keyword{datasets}
